[#update] 체력, 경험치 인터페이스 추가
최대 체력 및 레벨별 경험치는 임시

+++ Assets/Scripts/Enemy.cs
@@ -7,7 +7,7 @@ public class Enemy : Entity
 {
     public float moveSpeed = 3f;
 
-    float damage = 1f;
+    int damage = 1;
     bool isInit = false;
     bool isAlive;
 
@@ -50,7 +50,7 @@ public class Enemy : Entity
         }
     }
 
-    public override void OnDamage(Entity from, float damage)
+    public override void OnDamage(Entity from, int damage)
     {
         if (!isAlive) return;
